Add support for algorithm and hardness picking
- Adds support to choose the algorithm and the 'hardness' level for the
kdf function. Hardness sets specific numbers for the opslimit and
memlimit. Quoting from libsodium: 'opslimit represents a maximum amount
of computations to perform. Raising this number will make the function
require more CPU cycles to compute a key' and 'memlimit is the maximum
amount of RAM that the function will use, in bytes.' The reason for this
is that depending on the scenario, one should be able to select the
'paranoid' level for the kdf instead of opting for the max number
(previous situation)